feat: consolidate HomePageHeader into Header component, create SettingsButton component
